Data.Type.Length.Util

Documentation

append' :: Length ns -> Length ms -> Length (ns ++ ms)

reverse' :: forall ns. Length ns -> Length (Reverse ns)

(>:) :: Length ns -> Proxy n -> Length (ns >: n)

data SnocLength :: [a] -> Type where

Constructors

SnocZ :: SnocLength '[]

SnocS :: SnocLength as -> Proxy a -> SnocLength (as >: a)

snocLengthHelp :: forall as bs. Length bs -> SnocLength bs -> Length as -> SnocLength (bs ++ as)

snocLength :: Length as -> SnocLength as

could this be unsafeCoerce?

snocLengthLength :: SnocLength as -> Length as

could just be unsafeCoerce lol

snocLengthReverse :: SnocLength as -> Length (Reverse as)

data MaxLength :: N -> [k] -> Type where

A MaxLength n as is a witness that the list as has a length of at most n.

Constructors

FLZ :: MaxLength n '[]

FLS :: !(MaxLength n as) -> MaxLength (S n) (a ': as)

data ExactLength :: N -> [k] -> Type where

An ExactLength n as is a witness that the list as has a length of exactly n.

Constructors

ELZ :: ExactLength Z '[]

ELS :: !(ExactLength n as) -> ExactLength (S n) (a ': as)

data Splitting :: N -> ([k] -> Type) -> [k] -> Type where

Constructors

Fewer :: !(MaxLength n as) -> !(f as) -> Splitting n f as

Split :: !(ExactLength (S n) as) -> !(f as) -> !(f bs) -> Splitting n f (as ++ bs)

splitting :: Nat n -> Prod f as -> Splitting n (Prod f) as

maxLength :: Nat n -> Length as -> Decision (MaxLength n as)

exactLength :: Nat n -> Length as -> Decision (ExactLength n as)