TensorOps.BLAS

Documentation

data BShape a

Constructors

BV !a

BM !a !a

type family Equals_1627637949 (a :: BShape k) (b :: BShape k) :: Bool where ...

Equations

Equals_1627637949 (BV a) (BV b) = (:==) a b

Equals_1627637949 (BM a a) (BM b b) = (:&&) ((:==) a b) ((:==) a b)

Equals_1627637949 (a :: BShape k) (b :: BShape k) = FalseSym0

type BVSym1 t = BV t

data BVSym0 l

Constructors

(KindOf (Apply BVSym0 arg) ~ KindOf (BVSym1 arg)) => BVSym0KindInference

type BMSym2 t t = BM t t

data BMSym1 l l

Constructors

(KindOf (Apply (BMSym1 l) arg) ~ KindOf (BMSym2 l arg)) => BMSym1KindInference

data BMSym0 l

Constructors

(KindOf (Apply BMSym0 arg) ~ KindOf (BMSym1 arg)) => BMSym0KindInference

type family Compare_1627638263 (a :: BShape a) (a :: BShape a) :: Ordering where ...

Equations

Compare_1627638263 (BV a_1627637936) (BV b_1627637938) = Apply (Apply (Apply FoldlSym0 ThenCmpSym0) EQSym0) (Apply (Apply (:$) (Apply (Apply CompareSym0 a_1627637936) b_1627637938)) '[])

Compare_1627638263 (BM a_1627637940 a_1627637942) (BM b_1627637944 b_1627637946) = Apply (Apply (Apply FoldlSym0 ThenCmpSym0) EQSym0) (Apply (Apply (:$) (Apply (Apply CompareSym0 a_1627637940) b_1627637944)) (Apply (Apply (:$) (Apply (Apply CompareSym0 a_1627637942) b_1627637946)) '[]))

Compare_1627638263 (BV _z_1627638243) (BM _z_1627638246 _z_1627638249) = LTSym0

Compare_1627638263 (BM _z_1627638252 _z_1627638255) (BV _z_1627638258) = GTSym0

type Compare_1627638263Sym2 t t = Compare_1627638263 t t

data Compare_1627638263Sym1 l l

Constructors

(KindOf (Apply (Compare_1627638263Sym1 l) arg) ~ KindOf (Compare_1627638263Sym2 l arg)) => Compare_1627638263Sym1KindInference

data Compare_1627638263Sym0 l

Constructors

(KindOf (Apply Compare_1627638263Sym0 arg) ~ KindOf (Compare_1627638263Sym1 arg)) => Compare_1627638263Sym0KindInference

type SBShape = (Sing :: BShape a -> Type)

type family BShapeDims (s :: BShape k) = (ks :: [k]) | ks -> s where ...

Equations

BShapeDims (BV x) = '[x]

BShapeDims (BM x y) = '[x, y]

type BShapeDimsSym1 t = BShapeDims t

data BShapeDimsSym0 l

Constructors

(KindOf (Apply BShapeDimsSym0 arg) ~ KindOf (BShapeDimsSym1 arg)) => BShapeDimsSym0KindInference

class NatKind k => BLAS b where

Minimal complete definition

axpy, dot, norm, ger, gemv, gemm, indexB, indexRowB, transp, iRowsB, iElemsB, bgenA, bgenRowsA, eye, zero, zipB, traceB

Associated Types

type ElemB b :: Type

Methods

axpy :: ElemB b -> b (BV n) -> Maybe (b (BV n)) -> b (BV n)

dot :: b (BV n) -> b (BV n) -> ElemB b

norm :: b (BV n) -> ElemB b

ger :: b (BV n) -> b (BV m) -> b (BM n m)

gemv :: ElemB b -> b (BM n m) -> b (BV m) -> Maybe (ElemB b, b (BV n)) -> b (BV n)

gemm :: ElemB b -> b (BM n o) -> b (BM o m) -> Maybe (ElemB b, b (BM n m)) -> b (BM n m)

indexB :: Prod (IndexN k) (BShapeDims s) -> b s -> ElemB b

indexRowB :: IndexN k n -> b (BM n m) -> b (BV m)

transp :: b (BM n m) -> b (BM m n)

iRowsB :: Applicative f => (IndexN k n -> b (BV m) -> f (b (BV o))) -> b (BM n m) -> f (b (BM n o))

iElemsB :: Applicative f => (Prod (IndexN k) (BShapeDims s) -> ElemB b -> f (ElemB b)) -> b s -> f (b s)

bgenA :: Applicative f => (Prod (IndexN k) (BShapeDims s) -> f (ElemB b)) -> f (b s)

bgenRowsA :: Applicative f => (IndexN k n -> f (b (BV m))) -> f (b (BM n m))

eye :: b (BM n n)

zero :: b s

zipB :: (ElemB b -> ElemB b -> ElemB b) -> b s -> b s -> b s

traceB :: b (BM n n) -> ElemB b

elemsB :: (Applicative f, BLAS b) => (ElemB b -> f (ElemB b)) -> b s -> f (b s)

bgen :: forall k b s. BLAS b => (Prod (IndexN k) (BShapeDims s) -> ElemB b) -> b s

bgenRows :: BLAS b => (IndexN k n -> b (BV m)) -> b (BM n m)

data BTensor :: (k -> Type -> Type) -> (BShape k -> Type) -> [k] -> Type where

Constructors

BTS :: !(ElemB b) -> BTensor v b '[]

BTV :: !(b (BV n)) -> BTensor v b '[n]

BTM :: !(b (BM n m)) -> BTensor v b '[n, m]

BTN :: !(v n (BTensor v b (o ': (m ': ns)))) -> BTensor v b (n ': (o ': (m ': ns)))

unScalar :: BTensor v b '[] -> ElemB b

unVector :: BTensor v b '[n] -> b (BV n)

dispatchBLAS :: forall b ms os ns v. (Floating (ElemB b), BLAS b) => MaxLength N1 ms -> MaxLength N1 os -> MaxLength N1 ns -> BTensor v b (ms ++ os) -> BTensor v b (Reverse os ++ ns) -> BTensor v b (ms ++ ns)

TODO: add RULES pragmas so that this can be done without checking lengths at runtime in the common case that the lengths are known at compile-time.

Also, totally forgot about matrix-scalar multiplication here, but there isn't really any way of making it work without a lot of empty cases. should probably handle one level up.

mapRows :: forall k v ns ms os b. (Vec v, BLAS b) => Sing ns -> Length os -> (BTensor v b ms -> BTensor v b os) -> BTensor v b (ns ++ ms) -> BTensor v b (ns ++ os)

bRows :: forall k v ns ms os b f. (Applicative f, Vec v, BLAS b) => Sing ns -> Length os -> (BTensor v b ms -> f (BTensor v b os)) -> BTensor v b (ns ++ ms) -> f (BTensor v b (ns ++ os))

mapIxRows :: forall k v ns ms os b. (Vec v, BLAS b) => Sing ns -> Length os -> (Prod (IndexN k) ns -> BTensor v b ms -> BTensor v b os) -> BTensor v b (ns ++ ms) -> BTensor v b (ns ++ os)

foldMapIxRows :: forall k v ns ms m b. (Vec v, Monoid m, BLAS b) => Sing ns -> (Prod (IndexN k) ns -> BTensor v b ms -> m) -> BTensor v b (ns ++ ms) -> m

bIxRows :: forall k v ns ms os b f. (Applicative f, Vec v, BLAS b) => Sing ns -> Length os -> (Prod (IndexN k) ns -> BTensor v b ms -> f (BTensor v b os)) -> BTensor v b (ns ++ ms) -> f (BTensor v b (ns ++ os))

indexRowBTensor :: forall k b v ns ms. (BLAS b, Vec v) => Prod (IndexN k) ns -> BTensor v b (ns ++ ms) -> BTensor v b ms

mapBTensorElems :: (Vec v, BLAS b) => (ElemB b -> ElemB b) -> BTensor v b ns -> BTensor v b ns

bTensorElems :: forall k v ns b f. (Applicative f, Vec v, BLAS b) => (ElemB b -> f (ElemB b)) -> BTensor v b ns -> f (BTensor v b ns)

ifoldMapBTensor :: forall k v ns m b. (Monoid m, Vec v, BLAS b) => (Prod (IndexN k) ns -> ElemB b -> m) -> BTensor v b ns -> m

bTensorIxElems :: forall k v ns b f. (Applicative f, Vec v, BLAS b) => (Prod (IndexN k) ns -> ElemB b -> f (ElemB b)) -> BTensor v b ns -> f (BTensor v b ns)

zipBTensorElems :: forall v b ns. (BLAS b, Nesting1 Sing Applicative v) => Sing ns -> (ElemB b -> ElemB b -> ElemB b) -> BTensor v b ns -> BTensor v b ns -> BTensor v b ns

mapBTM :: forall k v ns n m ms b. (Vec v, BLAS b) => Length ns -> Length ms -> (b (BM n m) -> BTensor v b ms) -> BTensor v b (ns ++ '[n, m]) -> BTensor v b (ns ++ ms)

foldMapBTM :: (Monoid a, Vec v, BLAS b) => Length ns -> (b (BM n m) -> a) -> BTensor v b (ns ++ '[n, m]) -> a

traverseBTM :: forall k v ns n m ms b f. (Applicative f, Vec v, BLAS b) => Length ns -> Length ms -> (b (BM n m) -> f (BTensor v b ms)) -> BTensor v b (ns ++ '[n, m]) -> f (BTensor v b (ns ++ ms))

imapBTM :: forall k v ns n m ms b. (Vec v, BLAS b) => Length ns -> Length ms -> (Prod (IndexN k) ns -> b (BM n m) -> BTensor v b ms) -> BTensor v b (ns ++ '[n, m]) -> BTensor v b (ns ++ ms)

ifoldMapBTM :: (Vec v, Monoid a, BLAS b) => Length ns -> (Prod (IndexN k) ns -> b (BM n m) -> a) -> BTensor v b (ns ++ '[n, m]) -> a

itraverseBTM :: forall k v ns n m ms b f. (Applicative f, Vec v, BLAS b) => Length ns -> Length ms -> (Prod (IndexN k) ns -> b (BM n m) -> f (BTensor v b ms)) -> BTensor v b (ns ++ '[n, m]) -> f (BTensor v b (ns ++ ms))

mapBase :: forall v b ns. Nesting1 Proxy Functor v => (ElemB b -> ElemB b) -> (forall n. b (BV n) -> b (BV n)) -> (forall n m. b (BM n m) -> b (BM n m)) -> BTensor v b ns -> BTensor v b ns

zipBase :: forall v b ns. Nesting1 Sing Applicative v => Sing ns -> (ElemB b -> ElemB b -> ElemB b) -> (forall n. b (BV n) -> b (BV n) -> b (BV n)) -> (forall n m. b (BM n m) -> b (BM n m) -> b (BM n m)) -> BTensor v b ns -> BTensor v b ns -> BTensor v b ns

genBTensorA :: forall k b v ns f. (Applicative f, BLAS b, Vec v) => Sing ns -> (Prod (IndexN k) ns -> f (ElemB b)) -> f (BTensor v b ns)

genBTensor :: forall k b v ns. (BLAS b, Vec v) => Sing ns -> (Prod (IndexN k) ns -> ElemB b) -> BTensor v b ns

indexBTensor :: forall k b v ns. (BLAS b, Vec v) => Prod (IndexN k) ns -> BTensor v b ns -> ElemB b

btn :: (BLAS b, Vec v) => Length ns -> v n (BTensor v b ns) -> BTensor v b (n ': ns)

gmulB :: forall k b v ms os ns. (Floating (ElemB b), SingI ns, BLAS b, Vec v, Nesting1 Proxy Functor v, Nesting1 Sing Applicative v) => Sing ms -> Length os -> Length ns -> BTensor v b (ms ++ os) -> BTensor v b (Reverse os ++ ns) -> BTensor v b (ms ++ ns)

General strategy:

We can only outsource to BLAS (using dispatchBLAS) in the case that os and ns have length 0 or 1. Anything else, fall back to the basic reverse-indexing method in Data.Nested.

If ms is length 2 or higher, "traverse down" to the length 0 or 1 tail...and then sum them up.

naiveGMul :: forall k b v ms os ns. (BLAS b, Vec v, Num (ElemB b), Nesting1 Proxy Functor v, Nesting1 Sing Applicative v, SingI ns) => Sing ms -> Length os -> Length ns -> BTensor v b (ms ++ os) -> BTensor v b (Reverse os ++ ns) -> BTensor v b (ms ++ ns)

gmulBLAS :: forall b ms os ns v. (Floating (ElemB b), BLAS b, Vec v, SingI ns, Nesting1 Proxy Functor v, Nesting1 Sing Applicative v) => Sing ms -> MaxLength N1 os -> MaxLength N1 ns -> BTensor v b (ms ++ os) -> BTensor v b (Reverse os ++ ns) -> BTensor v b (ms ++ ns)